refactor: FilesEditor refactor to avoid super class
