+++ frontend/src/workspaces/features/FilesEditor/buildFileTree.ts
@@ -1,0 +1,22 @@
+import { AugmentedFile, FileNode } from "./types";
+
+export const buildFileTree = (flatNodes: AugmentedFile[]): FileNode[] => {
+  const nodeMap = new Map<string, FileNode>();
+
+  flatNodes.forEach((flatNode) => {
+    nodeMap.set(flatNode.id, { ...flatNode, children: [] });
+  });
+
+  flatNodes.forEach((flatNode) => {
+    if (flatNode.parentId) {
+      const parentNode = nodeMap.get(flatNode.parentId);
+      parentNode?.children.push(nodeMap.get(flatNode.id)!);
+    }
+  });
+
+  nodeMap.forEach((node) => {
+    node.children.sort((a, b) => a.name.localeCompare(b.name));
+  });
+
+  return Array.from(nodeMap.values());
+};
